todo-app/requirement.txt
Flask
Flask-GraphQL
Flask-SQLAlchemy
graphene
graphene-sqlalchemy
python-keycloak
